fix: 11 audio crash/stability fixes for recording and playback
Critical fixes for crash on recording stop and audio screeching:

Thread Safety:
- Added try_lock for m_channelMutex in processAudio (graph rebuild race)
- Removed inner lock_guard from processAudioSingleThreaded (deadlock)
- Refactored getMaxTimelineExtent to use PlaylistModel (UI use-after-free)
- Added global try-catch in processAudio (exception safety)
- Used getChannelsSnapshot() in MixerView::refreshChannels (UI race)

Buffer Bounds:
- Added break when source exhausted in resampling loop (source overrun)
- Clamped framesToProcess when bufferOffset > 0 (output overrun)
- Clamped clipFrames in waveform drawing (UI bounds)
- Added direct bounds check in AudioBufferData::getSample (vector safety)

Sample Rate:
- Replaced timing-based rate detection with m_inputSampleRate from driver
- Added setInputSampleRate to onStreamRestore for ASIO driver changes

+++ NomadAudio/include/AudioGraph.h
@@ -22,7 +22,7 @@ struct ClipRenderState {
     const float* audioData{nullptr};    // Interleaved stereo (engine format)
     uint64_t startSample{0};            // Absolute project sample (engine rate)
     uint64_t endSample{0};              // Exclusive end
-    uint64_t sampleOffset{0};           // Offset into audioData in frames
+    double sampleOffset{0.0};           // Offset into audioData in frames (double for sub-sample precision)
     uint64_t totalFrames{0};            // Bounds for audioData to guard OOB
     double sourceSampleRate{48000.0};   // Original clip sample rate
     float gain{1.0f};
